Fix error caused by train calling at station with no arrivial time
When searching form Redhill to London Bridge in the morning an API
result is returned for a service that calls at London Bridge without
having any arrival or departure time. I *think* this is showing up
because the service goes through London Bridge without stopping. Filter
services to ones with an arrival time to prevent this error.